Faster warpy - interpolation of transformation field (#25)
* Remove some commented code
* Adds a way to downscale and interpolate the transformation field.
* Adds downsampling transformation field as properties in the GUI of ImageCombinerWarpy
* Puts properties of RealTransformInterpolation as final, removes setters
* Fix https://github.com/BIOP/qupath-extension-warpy/issues/24 by adding +1

* Various CI modifications for QuPath 0.6 and Java 21
* Removes demo classes in test - seems to create issues with CI

In most cases, the transformation from one image to another does not vary very abruptly locally. Computing the exact transformation for each pixel is costly and mainly unnecessary in the image combiner warpy. This commit offers an option to increase the performance of the computation of the combined image by interpolating the transformation field on a grid, whose spacing is defined in pixels and chosen by the user.

A new flag, 'downsample', is present in the GUI to activate this speed-up. A new field is also present to specify the spacing of the grid (in pixels) where the transformation is computed 'exactly'. A reasonable value is 128 pixels.

All the modifications are backward compatible, and opening old projects will not lead to any issues. There's a good chance that these modifications are also 'forward compatible' and maintain the old behavior, but this has not been tested.

A few tests showed a speed-up of about a factor of 5 by interpolating the transform every 128 pixels and exported 9 1000 x 1000 pixel images (20s vs 1m20s) on a local bio-formats VSI slide. Note that this speed up may be visible only if there's a spline transformation. Affine or series of affine transformation may not go faster, and the interpolation may lead to slower computation time.

To achieve the speed-up, the transformation is computed on the grid ideally one time (or a few times if there are concurrent accesses). Note that the speed-up thus uses a bit more memory to store the raster transformation field at low resolution. If the downsampling is low (2 times), there's a chance this will cause a lot of trouble memory-wise (2 doubles per pixel). So the idea is really to get a high downsampling number (128 times): the memory cost becomes very low. For instance, a 128k x 128k image server will need to store 1k x 1k, which is about 1 million times 2 doubles (x and y), thus about 16 MB of RAM. Not negligible, but not terrible either for a 5 times speed-up.